Kalman_EDA_prediction_ICIC data
Here is the code for the project of Kalman filter to do the prediction of stock data from ICIC bank for 6months of 2022. It is the basis that I learned in the data science course. I have used my knowledge of Exploratory Data Analysis and made a prediction for the next 30 days in the code.   
I would love to get feedback and responses that can help me grow and develop more.